mercury/requirements.txt
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.4.0
certifi==2024.7.4
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
distro==1.9.0
fonttools==4.53.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
kiwisolver==1.4.5
Markdown==3.6
matplotlib==3.9.1
numpy==2.0.1
openai==1.37.0
packaging==24.1
pefile==2023.2.7
pillow==10.4.0
pydantic==2.8.2
pydantic_core==2.20.1
pyinstaller==6.9.0
pyinstaller-hooks-contrib==2024.7
pyparsing==3.1.2
PySide6==6.7.2
PySide6_Addons==6.7.2
PySide6_Essentials==6.7.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pywin32-ctypes==0.2.2
setuptools==71.1.0
shiboken6==6.7.2
six==1.16.0
sniffio==1.3.1
tqdm==4.66.4
typing_extensions==4.12.2
wheel==0.43.0
wordcloud==1.9.3
